requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
altair==5.4.1
annotated-types==0.7.0
anyio==4.4.0
async-timeout==4.0.3
attrs==23.2.0
blinker==1.8.2
cachetools==5.5.0
certifi==2024.7.4
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
comtypes==1.4.8
dataclasses-json==0.6.7
distro==1.9.0
dnspython==2.6.1
email_validator==2.2.0
exceptiongroup==1.2.2
fastapi==0.111.1
fastapi-cli==0.0.4
filelock==3.16.1
frozenlist==1.4.1
fsspec==2024.10.0
gitdb==4.0.11
GitPython==3.1.43
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
huggingface-hub==0.26.2
idna==3.7
Jinja2==3.1.4
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
langchain==0.2.11
langchain-community==0.2.10
langchain-core==0.2.24
langchain-openai==0.1.19
langchain-pinecone==0.1.3
langchain-text-splitters==0.2.2
langsmith==0.1.93
lxml==5.3.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.21.3
mdurl==0.1.2
mpmath==1.3.0
multidict==6.0.5
mypy-extensions==1.0.0
narwhals==1.11.1
networkx==3.4.2
numpy==1.26.4
openai==1.37.1
orjson==3.10.6
packaging==24.1
pandas==2.2.3
pillow==10.4.0
pinecone-client==5.0.0
pinecone-plugin-inference==1.0.2
pinecone-plugin-interface==0.0.7
protobuf==5.28.3
pyarrow==18.0.0
pycparser==2.22
pydantic==2.8.2
pydantic_core==2.20.1
pydeck==0.9.1
Pygments==2.18.0
PyMuPDF==1.24.9
PyMuPDFb==1.24.9
pypiwin32==223
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.0.1
python-multipart==0.0.9
pyttsx3==2.98
pytz==2024.2
pywin32==308
PyYAML==6.0.1
referencing==0.35.1
regex==2024.7.24
requests==2.32.3
rich==13.7.1
rpds-py==0.20.0
safetensors==0.4.5
shellingham==1.5.4
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
soundfile==0.12.1
SQLAlchemy==2.0.31
starlette==0.37.2
streamlit==1.39.0
sympy==1.13.1
tenacity==8.5.0
tiktoken==0.7.0
tokenizers==0.20.1
toml==0.10.2
torch==2.5.0
tornado==6.4.1
tqdm==4.66.4
transformers==4.46.0
typer==0.12.3
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.2
uvicorn==0.30.3
watchdog==5.0.3
watchfiles==0.22.0
websockets==12.0
yarl==1.9.4
